Fix bug in Dialects.Validate()
Validate() would always search each dialect. Thus if the last dialect
searched did not know about a message, it caused err to be set to
ErrUnknownMessage and thus cause validation to fail. Similarly, if a
dialect did know about a message, but was not the last dialect in
Dialects, Validate would ignore the metadata from that dialect and
keep searching through the last dialect. This bug in effect made
Validate() use only the last dialect.

This commit updates Validate() to quit searching if a dialect is aware
of a message.

Ideally, multiple dialects could be tried in the case where more than
one dialect has a match. This commit adds a TODO to cover this case.