Add option to FrameStream to return invalid frames
This updates FrameStream and adds an argument to NewFrameStream() to
allow returning invalid frames from readFrame(). In the case of
returning an invalid frame, parsing continues as if an invalid frame
were not returned. That is the buffer is shifted only by one byte and
the parsing continues from there on the next call to readFrame().

This is primarily meant for the case where FrameStream is being used
in a proxy. We want to pass on possibly unknown messages (i.e., CRC
seed is missing) as the other end will most likely validate the frame
again. FrameStream still uses the known messages to stay synchronized
in the stream.